Move the api calls to pet profile instead of page
